backend/requirements.txt
python-dotenv
fastapi
uvicorn
celery
redis
geopy
python-multipart
emails
requests
pytz
email-validator>=2.0.0